requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.8.1
billiard==3.6.4.0
celery==5.2.7
certifi
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.10
django-autoslug==1.9.8
django-browser-reload==1.6.0
django-ckeditor==6.4.2
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-environ==0.4.5
django-js-asset==2.0.0
django-paypal==2.0
django-tagify2==1.0.19
django-tailwind==3.3.0
django-widget-tweaks==1.4.12
filelock==3.0.12
idna==3.3
kombu==5.2.4
mysqlclient==2.2.4
Pillow~=10.2.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2==2.9.9
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python_http_client==3.3.2
pytz==2022.1
requests
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
stripe==3.5.0
toml==0.10.2
tornado==6.1
typing_extensions==4.11.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.2.0
xmltodict==0.12.0

pip~=24.0
attrs~=23.2.0
cryptography~=42.0.5
Jinja2~=3.1.3
setuptools~=69.2.0
packaging~=23.2
cffi~=1.16.0
numpy~=1.26.3
ipython
python-dateutil~=2.8.2
Pygments~=2.17.2
psutil~=5.9.8
anyio~=4.3.0
selenium